+++ src/routes/dashboard/[id]/+page.svelte
@@ -10,12 +10,13 @@
 	import type { SubmitFunction } from '@sveltejs/kit';
 	import { intlFormatDistance } from 'date-fns';
 	import { _ } from 'svelte-i18n';
-
+	import QRCode from 'qrcode';
 	import type { MouseEventHandler } from 'svelte/elements';
-	import { fly } from 'svelte/transition';
+	import { fade, fly } from 'svelte/transition';
 	import Chart from '$lib/ui/chart.svelte';
 	import { browser } from '$app/environment';
 	import { page } from '$app/stores';
+	import { outside } from '$lib/utils/outside';
 
 	let { data, form } = $props();
 	let active = $state(!data.snapp.disabled);
@@ -84,14 +85,107 @@
 			return;
 		}
 
-		if (idx && navigator.clipboard)
-			await navigator.clipboard.writeText($page.url.origin + '/' + idx);
+		if (idx && navigator.clipboard) await navigator.clipboard.writeText(fullUrlToSnapp);
 		toast.info($_('snapps.helpers.copied-to-clipboard'));
 	};
+
+	let show_zoomed_qrcode = $state(false);
+	let handle_show_zoomed_qrcode: MouseEventHandler<HTMLButtonElement> = (e) => {
+		e.preventDefault();
+		e.stopPropagation();
+		show_zoomed_qrcode = !show_zoomed_qrcode;
+		console.log(show_zoomed_qrcode);
+	};
+
+	let fullUrlToSnapp = $derived($page.url.origin + '/' + data.snapp.shortcode);
+
+	let can_share = $derived(browser === true && window?.navigator?.share);
 </script>
 
 <form action="?/disable" id="disable-snapp" hidden method="post" use:enhance={enhanceAction}></form>
 
+{#if show_zoomed_qrcode === true}
+	<div id="zoomed" class="fixed inset-0 z-[100] flex flex-col bg-neutral-950/50" transition:fade>
+		<div
+			use:outside={() => {
+				show_zoomed_qrcode = false;
+			}}
+			class="z-30 m-auto flex rounded w-full max-w-sm flex-col gap-4 overflow-y-scroll text-balance border border-slate-500/50 bg-neutral-50 p-4 leading-relaxed text-neutral-950 transition-colors dark:bg-neutral-950 dark:text-neutral-50"
+			transition:fly={{ y: 25 }}
+		>
+			{#await QRCode.toDataURL(fullUrlToSnapp, { width: 350, margin: 2 })}
+				<Icon ph="spinner" />
+			{:then generated}
+				<div class="flex gap-4 flex-col">
+					<Card css={{ card: 'p-2' }}>
+						<img
+							src={generated}
+							class="w-full h-full aspect-square"
+							alt="QRCode: {data.snapp.shortcode}"
+						/>
+					</Card>
+					<Card css={{ card: 'items-center text-balance w-full leading-relaxed gap-4 p-2' }}>
+						<div class="flex flex-col md:flex-row w-full h-full gap-4">
+							<div class="flex w-full h-full gap-2 flex-col items-center justify-center">
+								<span class="w-full text-sm font-bold">QR Code</span>
+								<div class="flex gap-4 flex-col justify-center w-full">
+									<button
+										onclick={async () => {
+											const a = document.createElement('a');
+											a.href = await QRCode.toDataURL(fullUrlToSnapp, {
+												width: 800,
+												margin: 4
+											});
+											a.title = 'QRCode: [' + data.snapp.shortcode + ']';
+											a.download = 'qrcode-' + data.snapp.shortcode + '.png';
+											a.click();
+										}}
+										class="flex h-10 flex-row w-full items-center gap-2 rounded border-none bg-slate-500/25 p-0 px-4 text-start font-semibold outline-none transition-all hover:bg-slate-500/50 focus:bg-slate-500/50 md:h-8"
+										><Icon ph="download" />
+										<small class="text-sm">{$_('globals.download')}</small></button
+									>
+									{#if can_share}
+										<button
+											onclick={async () => {
+												const filename = data.snapp.shortcode + '.png';
+												const dataUri = await QRCode.toDataURL(fullUrlToSnapp);
+												function dataUriToBlob(dataURI: string, dataTYPE: string) {
+													var binary = atob(dataURI.split(',')[1]),
+														array = [];
+													for (var i = 0; i < binary.length; i++) array.push(binary.charCodeAt(i));
+													return new Blob([new Uint8Array(array)], { type: dataTYPE });
+												}
+
+												const _data = {
+													files: [
+														new File([dataUriToBlob(dataUri, 'image/png')], filename, {
+															type: 'image/png'
+														})
+													],
+													title: filename,
+													text: fullUrlToSnapp
+												};
+												try {
+													if (!navigator.canShare(_data)) return;
+
+													await navigator.share(_data);
+												} catch (err) {
+													console.log(err);
+												}
+											}}
+											class="flex h-10 flex-row w-full max-w-none items-center gap-2 rounded border-none bg-slate-500/25 p-0 px-4 text-start font-semibold outline-none transition-all hover:bg-slate-500/50 focus:bg-slate-500/50 md:h-8"
+											><Icon ph="share-network" />
+											<small class="text-sm">{$_('globals.share')}</small></button
+										>{/if}
+								</div>
+							</div>
+						</div>
+					</Card>
+				</div>
+			{/await}
+		</div>
+	</div>
+{/if}
 <div class={cn('flex w-full flex-col p-4 pb-8', show_tab === 'notes' ? 'h-full' : '')}>
 	<div class="mx-auto flex h-full w-full max-w-5xl flex-col gap-4">
 		<a class="flex gap-2 font-semibold uppercase tracking-wider" href="/dashboard"
@@ -157,6 +251,12 @@
 					show_tab === 'notes' ? 'bg-slate-500/50' : 'bg-slate-500/25 '
 				)}>{$_('snapps.fields.notes')}</button
 			>
+			<button
+				class={cn(
+					'flex h-10 w-full max-w-max items-center gap-2 rounded p-2 text-sm bg-slate-500/25 leading-none transition-all hover:bg-slate-500/50 focus:bg-slate-500/50 md:h-8'
+				)}
+				onclick={handle_show_zoomed_qrcode}>QR Code</button
+			>
 			<a
 				href="/dashboard/{data.snapp.id}/edit"
 				class={cn(
@@ -168,12 +268,18 @@
 		</Card>
 		{#if show_tab === 'details'}
 			<div class="flex w-full" in:fly={{ y: 25 }}>
-				<Card css={{ card: 'gap-4' }}>
+				<Card css={{ card: 'gap-4 w-full' }}>
 					<div class="flex w-full flex-col gap-4 lg:flex-row">
-						<Card css={{ card: 'items-center text-balance w-full leading-relaxed gap-4 p-2' }}>
+						<Card
+							css={{
+								card: 'items-center text-balance max-w-lg w-full leading-relaxed gap-4 p-2'
+							}}
+						>
 							<Switch
 								name="status"
-								label="{$_('snapps.fields.status')}: {active === true ? $_("globals.active") : $_("globals.disabled")}"
+								label="{$_('snapps.fields.status')}: {active === true
+									? $_('globals.active')
+									: $_('globals.disabled')}"
 								helper={$_('snapps.helpers.disable-text-1')}
 								actions={{ toggle: handle_disable }}
 								bind:value={active}
@@ -181,29 +287,29 @@
 							<small class="w-full">{$_('snapps.helpers.disable-text-2')}</small>
 						</Card>
 						<div class="flex w-full flex-col gap-4">
-							<Card css={{ card: 'p-2 gap-4 w-full' }}>
+							<Card css={{ card: 'p-2 gap-4' }}>
 								<Input
 									icons={{ left: 'globe' }}
 									name="original_url"
 									disabled
 									label={$_('snapps.fields.original-url')}
-									bind:value={data.snapp.original_url}
+									value={data.snapp.original_url}
 								/>
 							</Card>
-							<Card css={{ card: 'p-2 gap-4' }}>
+							<Card css={{ card: 'p-2 gap-4 max-w-none w-full' }}>
 								<Input
 									icons={{ left: 'link-simple-horizontal' }}
 									name="shortcode"
 									disabled
 									label={$_('snapps.fields.shortcode')}
-									bind:value={data.snapp.shortcode}
+									value={data.snapp.shortcode}
 								/>
 							</Card>
 						</div>
 					</div>
 					<div class="flex w-full flex-col gap-4 lg:flex-row">
 						<Card
-							css={{ card: 'items-center text-balance w-fullleading-relaxed gap-4 p-2 h-full' }}
+							css={{ card: 'items-center text-balance w-full leading-relaxed gap-4 p-2 h-full' }}
 						>
 							<Input
 								icons={{ left: 'lock-laminated' }}
@@ -216,7 +322,7 @@
 									: $_('snapps.helpers.secret')}
 							/>
 						</Card>
-						<Card css={{ card: 'p-2 gap-4 flex-row w-full lg:max-w-[calc(50%_-_.75rem)]' }}>
+						<Card css={{ card: 'p-2 gap-4 flex-row w-full' }}>
 							<Input
 								name="hit"
 								disabled
@@ -224,7 +330,6 @@
 								css={{ input: 'capitalize' }}
 								value={data.snapp.hit}
 							/>
-
 							<Input
 								name="used"
 								disabled
@@ -270,7 +375,7 @@
 								value={getRelativeDate(data.snapp.created)}
 							/>
 						</Card>
-						<Card css={{ card: 'p-2 gap-4 flex-row w-full lg:max-w-[calc(50%_-_.75rem)]' }}>
+						<Card css={{ card: 'p-2 gap-4 flex-row w-full' }}>
 							<Input
 								icons={{ left: 'alarm' }}
 								name="expiration"